+++ .gitignore
@@ -131,3 +131,6 @@ dmypy.json
 
 # PyCharm IDE:
 .idea
+
+# MacOS Finder
+.DS_Store
