+++ lastfm/charts.py
@@ -6,9 +6,8 @@ from redbot.core import commands
 from redbot.core.utils import AsyncIter
 
 from .abc import MixinMeta
-from .errors import *
+from .exceptions import *
 from .fmmixin import fm
-from .utils import *
 
 NO_IMAGE_PLACEHOLDER = (
     "https://lastfm.freetls.fastly.net/i/u/300x300/2a96cbd8b46e442fc41c2b86b821562f.png"
@@ -32,8 +31,8 @@ class ChartMixin(MixinMeta):
     async def chart(self, ctx, *args):
         """Visual chart of your top albums or artists."""
         conf = await self.config.user(ctx.author).all()
-        check_if_logged_in(conf)
-        arguments = parse_chart_arguments(args)
+        self.check_if_logged_in(conf)
+        arguments = self.parse_chart_arguments(args)
         if arguments["width"] + arguments["height"] > 31:  # TODO: Figure out a reasonable value.
             return await ctx.send(
                 "Size is too big! Chart `width` + `height` total must not exceed `31`"
@@ -65,7 +64,7 @@ class ChartMixin(MixinMeta):
                         self.chart_data[album["image"][3]["#text"]] = chart_img
                     chart.append(
                         (
-                            f"{plays} {format_plays(plays)}\n{name} - {artist}",
+                            f"{plays} {self.format_plays(plays)}\n{name} - {artist}",
                             chart_img,
                         )
                     )
@@ -95,7 +94,7 @@ class ChartMixin(MixinMeta):
                         self.chart_data[scraped_images[i]] = chart_img
                     chart.append(
                         (
-                            f"{plays} {format_plays(plays)}\n{name}",
+                            f"{plays} {self.format_plays(plays)}\n{name}",
                             chart_img,
                         )
                     )
@@ -137,7 +136,7 @@ class ChartMixin(MixinMeta):
         u = conf["lastfm_username"]
         try:
             await ctx.send(
-                f"`{u} - {humanized_period(arguments['period'])} - {arguments['width']}x{arguments['height']} {chart_type} chart`",
+                f"`{u} - {self.humanized_period(arguments['period'])} - {arguments['width']}x{arguments['height']} {chart_type} chart`",
                 file=img,
             )
         except discord.HTTPException:
