+++ lastfm/exceptions.py
@@ -1,11 +1,14 @@
 class LastFMError(Exception):
     pass
 
+
 class NotLoggedInError(LastFMError):
     pass
 
+
 class NeedToReauthorizeError(LastFMError):
     pass
+
 
 class SilentDeAuthorizedError(LastFMError):
     pass
