+++ lastfm/fmmixin.py
@@ -1,6 +1,6 @@
 from redbot.core import commands
 
-from .utils import tokencheck
+from .utils.tokencheck import tokencheck
 
 
 @commands.check(tokencheck)
@@ -15,7 +15,3 @@ class FMMixin:
     """This is mostly here to easily mess with things..."""
 
     c = fm
-
-
-async def red_delete_data_for_user(self, *, requester, user_id):
-    await self.config.user_from_id(user_id).clear()
