+++ lastfm/top.py
@@ -3,9 +3,8 @@ from redbot.core.utils.chat_formatting import escape
 from redbot.core.utils.menus import DEFAULT_CONTROLS, menu
 
 from .abc import MixinMeta
-from .errors import *
+from .exceptions import *
 from .fmmixin import fm
-from .utils import *
 
 
 class TopMixin(MixinMeta):
@@ -15,9 +14,9 @@ class TopMixin(MixinMeta):
     async def topartists(self, ctx, *args):
         """Most listened artists."""
         conf = await self.config.user(ctx.author).all()
-        check_if_logged_in(conf)
+        self.check_if_logged_in(conf)
         async with ctx.typing():
-            arguments = parse_arguments(args)
+            arguments = self.parse_arguments(args)
             data = await self.api_request(
                 ctx,
                 {
@@ -37,21 +36,19 @@ class TopMixin(MixinMeta):
             for i, artist in enumerate(artists, start=1):
                 name = escape(artist["name"], formatting=True)
                 plays = artist["playcount"]
-                rows.append(f"`#{i:2}` **{plays}** {format_plays(plays)} — **{name}**")
+                rows.append(f"`#{i:2}` **{plays}** {self.format_plays(plays)} — **{name}**")
 
             image_url = await self.scrape_artist_image(artists[0]["name"], ctx)
-            # image_colour = await color_from_image_url(image_url)
 
             content = discord.Embed(color=await self.bot.get_embed_color(ctx.channel))
-            # content.colour = int(image_colour, 16)
             content.set_thumbnail(url=image_url)
             content.set_footer(text=f"Total unique artists: {user_attr['total']}")
             content.set_author(
-                name=f"{user_attr['user']} — {humanized_period(arguments['period']).capitalize()} top artists",
+                name=f"{user_attr['user']} — {self.humanized_period(arguments['period']).capitalize()} top artists",
                 icon_url=ctx.message.author.avatar_url,
             )
 
-        pages = await create_pages(content, rows)
+        pages = await self.create_pages(content, rows)
         if len(pages) > 1:
             await menu(ctx, pages[:15], DEFAULT_CONTROLS)
         else:
@@ -61,8 +58,8 @@ class TopMixin(MixinMeta):
     async def topalbums(self, ctx, *args):
         """Most listened albums."""
         conf = await self.config.user(ctx.author).all()
-        check_if_logged_in(conf)
-        arguments = parse_arguments(args)
+        self.check_if_logged_in(conf)
+        arguments = self.parse_arguments(args)
         data = await self.api_request(
             ctx,
             {
@@ -84,23 +81,20 @@ class TopMixin(MixinMeta):
             artist_name = escape(album["artist"]["name"], formatting=True)
             plays = album["playcount"]
             rows.append(
-                f"`#{i:2}` **{plays}** {format_plays(plays)} — **{artist_name}** — ***{name}***"
+                f"`#{i:2}` **{plays}** {self.format_plays(plays)} — **{artist_name}** — ***{name}***"
             )
 
         image_url = albums[0]["image"][-1]["#text"]
-        # image_url_small = albums[0]['image'][1]['#text']
-        # image_colour = await color_from_image_url(image_url_small)
 
         content = discord.Embed(color=await self.bot.get_embed_color(ctx.channel))
-        # content.colour = int(image_colour, 16)
         content.set_thumbnail(url=image_url)
         content.set_footer(text=f"Total unique albums: {user_attr['total']}")
         content.set_author(
-            name=f"{user_attr['user']} — {humanized_period(arguments['period']).capitalize()} top albums",
+            name=f"{user_attr['user']} — {self.humanized_period(arguments['period']).capitalize()} top albums",
             icon_url=ctx.message.author.avatar_url,
         )
 
-        pages = await create_pages(content, rows)
+        pages = await self.create_pages(content, rows)
         if len(pages) > 1:
             await menu(ctx, pages[:15], DEFAULT_CONTROLS)
         else:
@@ -110,9 +104,9 @@ class TopMixin(MixinMeta):
     async def toptracks(self, ctx, *args):
         """Most listened tracks."""
         conf = await self.config.user(ctx.author).all()
-        check_if_logged_in(conf)
+        self.check_if_logged_in(conf)
         async with ctx.typing():
-            arguments = parse_arguments(args)
+            arguments = self.parse_arguments(args)
             data = await self.api_request(
                 ctx,
                 {
@@ -134,7 +128,7 @@ class TopMixin(MixinMeta):
                 artist_name = escape(track["artist"]["name"], formatting=True)
                 plays = track["playcount"]
                 rows.append(
-                    f"`#{i:2}` **{plays}** {format_plays(plays)} — **{artist_name}** — ***{name}***"
+                    f"`#{i:2}` **{plays}** {self.format_plays(plays)} — **{artist_name}** — ***{name}***"
                 )
             trackdata = await self.api_request(
                 ctx,
@@ -160,11 +154,11 @@ class TopMixin(MixinMeta):
 
             content.set_footer(text=f"Total unique tracks: {user_attr['total']}")
             content.set_author(
-                name=f"{user_attr['user']} — {humanized_period(arguments['period']).capitalize()} top tracks",
+                name=f"{user_attr['user']} — {self.humanized_period(arguments['period']).capitalize()} top tracks",
                 icon_url=ctx.message.author.avatar_url,
             )
 
-            pages = await create_pages(content, rows)
+            pages = await self.create_pages(content, rows)
             if len(pages) > 1:
                 await menu(ctx, pages[:15], DEFAULT_CONTROLS)
             else:
